Version 1.0.1 - Docfix and QR method separate
