+++ bankid/__init__.py
@@ -23,8 +23,7 @@ from bankid.__version__ import __version__, version
 from bankid.certutils import create_bankid_test_server_cert_and_key
 from bankid.syncclient import BankIDClient
 from bankid.asyncclient import BankIDAsyncClient
-from bankid.baseclient import generate_qr_code_content
-
+from bankid.qr import generate_qr_code_content
 
 __all__ = [
     "BankIDClient",
